% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIDIMS-package.R
\docType{package}
\name{rIDIMS-package}
\alias{rIDIMS}
\alias{rIDIMS-package}
\title{rIDIMS: An R Package for Processing Intermittent and Direct Injection Mass Spectrometry Data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/felipe-mansoldo/IDIMS}
  \item Report bugs at \url{https://github.com/felipe-mansoldo/IDIMS/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
