% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_IDIMS.R, R/start_rIDIMS.R
\name{start_rIDIMS}
\alias{start_rIDIMS}
\title{start_rIDIMS}
\usage{
start_rIDIMS(debug_app = TRUE, QC_app = TRUE)

start_rIDIMS(debug_app = TRUE, QC_app = TRUE)
}
\value{
null

null
}
\description{
start_rIDIMS

start_rIDIMS
}
