% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIDIMS-package.R
\docType{package}
\name{rIDIMS-package}
\alias{rIDIMS}
\alias{rIDIMS-package}
\title{rIDIMS: An R Package for Processing Intermittent and Direct Injection Mass Spectrometry Data}
\description{
rIDIMS offers a simple and fast workflow to process two-dimensional mass spec data by selecting high-quality scans and features based on mass spectrum intensity, data alignment, filtering reproducible ions along sample replicates, removing background ions, and selecting ions that are more representative per class.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BioinovarLab/rIDIMS}
  \item \url{https://bioinovarlab.github.io/rIDIMS/}
  \item Report bugs at \url{https://github.com/BioinovarLab/rIDIMS/issues}
}

}
\author{
\strong{Maintainer}: Felipe Mansoldo \email{mansoldo@micro.ufrj.br}

Authors:
\itemize{
  \item Iasmim Lopes \email{minklopes@gmail.com}
}

}
\keyword{internal}
